ledger_service/requirements.txt
# Dependencias de Python para el Ledger Service.

fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
cassandra-driver
prometheus-client